+++ geode-core/src/main/java/org/apache/geode/management/internal/api/GeodeClusterManagementServiceFactory.java
@@ -36,19 +36,19 @@ import org.apache.geode.internal.logging.LogService;
 import org.apache.geode.management.api.ClusterManagementService;
 import org.apache.geode.management.client.ClusterManagementServiceProvider;
 import org.apache.geode.management.internal.ClusterManagementClient;
+import org.apache.geode.management.internal.JavaClientClusterManagementFactory;
 import org.apache.geode.management.internal.cli.domain.MemberInformation;
 import org.apache.geode.management.internal.cli.functions.GetMemberInformationFunction;
-import org.apache.geode.management.spi.ClusterManagementServiceProviderFactory;
-import org.apache.geode.management.spi.JavaClientClusterManagementProviderFactory;
+import org.apache.geode.management.spi.ClusterManagementServiceFactory;
 
 /**
- * An implementation of {@link ClusterManagementServiceProviderFactory} which can be used in any
+ * An implementation of {@link ClusterManagementServiceFactory} which can be used in any
  * context where Geode is running (client, server or locator). It will attempt to determine the
  * address of the {@code ClusterManagementService} when using the {@code create()} call. Otherwise
  * an explicit can also be used.
  */
-public class GeodeClusterManagementServiceProviderFactory extends
-    JavaClientClusterManagementProviderFactory {
+public class GeodeClusterManagementServiceFactory extends
+    JavaClientClusterManagementFactory {
 
   @Immutable
   private static final GetMemberInformationFunction MEMBER_INFORMATION_FUNCTION =
