+++ geode-management/src/main/resources/META-INF/services/org.apache.geode.management.spi.ClusterManagementServiceFactory
@@ -13,4 +13,5 @@
 # the License.
 #
 
-org.apache.geode.management.internal.api.GeodeClusterManagementServiceProviderFactory
+
+org.apache.geode.management.internal.JavaClientClusterManagementFactory
